requirements.txt
Flask==1.1.1
slackclient==2.5.0
gspread==3.5.0
